@@ -6,7 +6,8 @@
         "illuminate/console": "^8.0 | ^6.0 | ^7.0 | ^5.6",
         "illuminate/support": "^8.0 | ^6.0 | ^7.0 | ^5.6",
         "lorisleiva/cron-translator": "^0.2.0",
-        "php": ">= 7.1"
+        "php": ">= 8.0",
+        "artarts36/ci-git-sender": "^0.1.1"
     },
     "require-dev": {
         "squizlabs/php_codesniffer": "3.5",
