@@ -6,7 +6,7 @@ use ArtARTs36\LaravelScheduleDocumentator\Contracts\DocumentContent;
 use ArtARTs36\LaravelScheduleDocumentator\Data\EventCollection;
 use ArtARTs36\LaravelScheduleDocumentator\Data\StringDocumentContent;
 use ArtARTs36\LaravelScheduleDocumentator\Exceptions\TemplateNotFound;
-use Illuminate\Contracts\Filesystem\Filesystem;
+use Illuminate\Filesystem\Filesystem;
 use Illuminate\View\Factory;
 
 abstract class TemplateDocumentator extends AbstractDocumentator
